🗺️ Atlas: [Split The Blob: Assembly]
Extracted the data transfer objects (`AssembledStatement`, `Constituent`, etc.) from the monolithic `src/semantic/assembly.rs` file into a new `src/semantic/assembly/model.rs` module, splitting the file into a clean data vs logic separation.

Updated imports in internal tests and modules to properly fetch from the new paths and maintained the public re-export (`pub use model::*`) to preserve compatibility. Added the proper structural finding log to `.jules/atlas.md`.
