--- plan.md
@@ -1,44 +1,0 @@
-1. **Refactoring the God Function `generate_expr` in `src/codegen.rs`:**
-   - I will extract logic from `generate_expr` in `src/codegen.rs` without returning `Option` or using `unreachable!()`.
-   - I will use `replace_with_git_merge_diff` to restructure `generate_expr`. I will create smaller helper functions and route to them, but the main match block in `generate_expr` will remain exhaustive.
-   - Wait, `generate_expr` already delegates to `generate_literal_string`, `generate_literal_number`, `generate_collection_array`. The "God Object" issue is probably related to the inline logic for `Some`, `None`, `Ok`, `Err`, `Try`, `Unwrap`. I will extract these into `generate_expr_some`, `generate_expr_none`, `generate_expr_ok`, `generate_expr_err`, `generate_expr_try`, `generate_expr_unwrap` returning `TokenStream`, and call them directly from the arms of the `match` block in `generate_expr`.
-   - Let's look at `generate_statement`. It delegates everything already. `AnalyzedStatement::Binding { ... } => generate_statement_binding(...)`, etc.
-
-Wait, if `generate_expr` is already delegating to helpers for most cases, what exactly was I supposed to refactor?
-Ah! Let's examine `generate_expr` again.
-
-```rust
-        AnalyzedExprKind::Some(inner) => {
-            let inner_tokens = generate_expr(inner);
-            quote! { Some(#inner_tokens) }
-        }
-```
-
-This inline logic can be extracted to:
-```rust
-fn generate_expr_some(inner: &AnalyzedExpr) -> TokenStream {
-    let inner_tokens = generate_expr(inner);
-    quote! { Some(#inner_tokens) }
-}
-```
-
-And in `generate_expr`:
-```rust
-        AnalyzedExprKind::Some(inner) => generate_expr_some(inner),
-```
-
-This makes the `match` a pure router.
-
-Let's also extract `Ok`, `Err`, `Try`, `Unwrap`.
-
-2. **Verify Changes:**
-   - Run `git diff`.
-
-3. **Run Validations:**
-   - `cargo fmt`, `cargo test`, `cargo clippy`.
-
-4. **Pre-commit Steps:**
-   - Pre-commit checks.
-
-5. **Submit:**
-   - Submit the change.
